requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
deepl==1.11.0
idna==3.4
inquirerpy==0.3.4
pfzy==0.3.4
prompt-toolkit==3.0.32
Pygments==2.13.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
urllib3==1.26.12
wcwidth==0.2.5
